# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "이화여대 창업 교육 LMS API",
        "title": "EarnLearning LMS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "earnlearning.com",
    "basePath": "/api",
    "paths": {
...
        "/admin/wallet/transfer": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "관리자용: 여러 사용자에게 일괄 송금",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "관리자 일괄 송금",
                "parameters": [
                    {
                        "description": "일괄 송금 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.AdminTransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/wallet": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "내 지갑 잔액 및 정보 조회",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "지갑 조회",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/wallet/ranking": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "전체 사용자 자산 랭킹 조회",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "자산 랭킹",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "조회 수",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/wallet/recipients": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "이름/이메일로 송금 대상 사용자 검색",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "송금 대상 검색",
                "parameters": [
                    {
                        "type": "string",
                        "description": "검색어",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/wallet/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "내 지갑 거래 내역 조회 (필터, 페이지네이션)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "거래 내역 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "페이지 번호",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "페이지 크기",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "거래 유형 필터",
                        "name": "tx_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "시작일 (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "종료일 (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/wallet/transfer": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "다른 사용자에게 송금",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "송금",
                "parameters": [
                    {
                        "description": "송금 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_interfaces_http_handler.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_INPUT"
                },
                "message": {
                    "type": "string",
                    "example": "잘못된 입력입니다"
                }
            }
        },
        "internal_interfaces_http_handler.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/internal_interfaces_http_handler.APIError"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "internal_interfaces_http_handler.AdminTransferRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 5000
                },
                "description": {
                    "type": "string",
                    "example": "보상금 지급"
                },
                "user_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "internal_interfaces_http_handler.TransferRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 1000
                },
                "description": {
                    "type": "string",
                    "example": "송금 메모"
                },
                "target_user_id": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}